+++ test/metabase/driver/clickhouse_test_utils.clj
@@ -56,12 +56,31 @@
                       " (1, true, true),"
                       " (2, false, true),"
                       " (3, true, false);")
+                 (str "CREATE TABLE `metabase_test`.`maps_test`"
+                      " (m Map(String, UInt64)) ENGINE = Memory;")
+                 (str "INSERT INTO `metabase_test`.`maps_test` VALUES"
+                      " ({'key1':1, 'key2':10}), ({'key1':2,'key2':20}), ({'key1':3,'key2':30});")
+                 ;; Used for testing that AggregateFunction columns are excluded,
+                 ;; while SimpleAggregateFunction columns are preserved
                  (str "CREATE TABLE `metabase_test`.`aggregate_functions_filter_test` ("
                       " idx UInt8, a AggregateFunction(uniq, String), lowest_value SimpleAggregateFunction(min, UInt8),"
                       " count SimpleAggregateFunction(sum, Int64)"
                       ") ENGINE Memory;")
                  (str "INSERT INTO `metabase_test`.`aggregate_functions_filter_test`"
-                      " (idx, lowest_value, count) VALUES (42, 144, 255255);")]]
+                      " (idx, lowest_value, count) VALUES (42, 144, 255255);")
+                 ;; Materialized views (testing .inner tables exclusion)
+                 (str "CREATE TABLE `metabase_test`.`wikistat` ("
+                      " `date` Date,"
+                      " `project` LowCardinality(String),"
+                      " `hits` UInt32"
+                      ") ENGINE = Memory;")
+                 (str "CREATE MATERIALIZED VIEW `metabase_test`.`wikistat_mv` ENGINE=Memory AS"
+                      " SELECT date, project, sum(hits) AS hits FROM `metabase_test`.`wikistat`"
+                      " GROUP BY date, project;")
+                 (str "INSERT INTO `metabase_test`.`wikistat` VALUES"
+                      " (now(), 'foo', 10),"
+                      " (now(), 'bar', 10),"
+                      " (now(), 'bar', 20);")]]
       (jdbc/execute! conn [sql]))))
 
 (defn do-with-metabase-test-db